Pull english strings out of javascript.
The account selector had the string "create new or select existing"
which was constructed in application.js and therefore couldn't be
localized.

This patch moves the string construction into the views so it can
be constructed with t().

I've used the key :select_existing which will need to be added to
the translation files.